Initial commit of LightUI variant
The `LightUI` variant preserves the Cobalt2 styling in the editor itself but uses a light colored theme for the VS Code user interface elements.

This will be preferred for anyone who likes the cobalt colors when editing code and text but finds the overall UI more difficult to use with a dark styling.

The light color choices have been borrowed from the `GitHub Light Default` theme. The license is compatible with this usage. https://github.com/primer/github-vscode-theme